Now the program runs and writes to disk. To check that the propagation is right

+++ include/utils/iParms.h
@@ -15,8 +15,8 @@
 		double		wDz;
 
 		InitialCond	cType;
-		int		parm1;
-		double		parm2;
+		double		parm1;
+		int		parm2;
 
 		FieldExpansion	fExp;
 		double		lSize;
